AG-NONE ag-charts-community: move chart/dom and chart/locale dirs
These are being move up one directory at the root of src/
Paths updated too.
