+++ packages/ag-charts-community/src/chart/chartContext.ts
@@ -8,9 +8,9 @@ import { AnnotationManager } from './annotation/annotationManager';
 import { AxisManager } from './axis/axisManager';
 import type { ChartService } from './chartService';
 import { DataService } from './data/dataService';
-import { DOMManager } from './dom/domManager';
-import { FocusIndicator } from './dom/focusIndicator';
-import { ProxyInteractionService } from './dom/proxyInteractionService';
+import { DOMManager } from '../dom/domManager';
+import { FocusIndicator } from '../dom/focusIndicator';
+import { ProxyInteractionService } from '../dom/proxyInteractionService';
 import { AnimationManager } from './interaction/animationManager';
 import { AriaAnnouncementService } from './interaction/ariaAnnouncementServices';
 import { ChartEventManager } from './interaction/chartEventManager';
@@ -27,7 +27,7 @@ import { TooltipManager } from './interaction/tooltipManager';
 import type { ZoomManager } from './interaction/zoomManager';
 import type { Keyboard } from './keyboard';
 import { LayoutManager } from './layout/layoutManager';
-import { LocaleManager } from './locale/localeManager';
+import { LocaleManager } from '../locale/localeManager';
 import { SeriesStateManager } from './series/seriesStateManager';
 import type { Tooltip } from './tooltip/tooltip';
 import { type UpdateCallback, UpdateService } from './updateService';
