+++ packages/ag-charts-community/src/chart/interaction/cursorManager.ts
@@ -1,5 +1,5 @@
 import { StateTracker } from '../../util/stateTracker';
-import type { DOMManager } from '../dom/domManager';
+import type { DOMManager } from '../../dom/domManager';
 
 export enum Cursor {
     Auto = 'auto',
