+++ packages/ag-charts-community/src/chart/interaction/interactionManager.ts
@@ -3,7 +3,7 @@ import { getWindow } from '../../util/dom';
 import { Logger } from '../../util/logger';
 import { partialAssign } from '../../util/object';
 import { isFiniteNumber } from '../../util/type-guards';
-import type { DOMManager } from '../dom/domManager';
+import type { DOMManager } from '../../dom/domManager';
 import { InteractionState, InteractionStateListener } from './interactionStateListener';
 import { type PreventableEvent, type Unpreventable, dispatchTypedEvent } from './preventableEvent';
