+++ packages/ag-charts-community/src/chart/overlay/chartOverlays.ts
@@ -1,6 +1,6 @@
 import { BaseProperties } from '../../util/properties';
 import { BOOLEAN, OBJECT, Validate } from '../../util/validation';
-import type { LocaleManager } from '../locale/localeManager';
+import type { LocaleManager } from '../../locale/localeManager';
 import { Overlay } from './overlay';
 
 export class ChartOverlays extends BaseProperties {
