+++ packages/ag-charts-community/src/chart/overlay/overlay.ts
@@ -3,7 +3,7 @@ import { createElement } from '../../util/dom';
 import { BaseProperties } from '../../util/properties';
 import { FUNCTION, STRING, Validate } from '../../util/validation';
 import type { AnimationManager } from '../interaction/animationManager';
-import type { LocaleManager } from '../locale/localeManager';
+import type { LocaleManager } from '../../locale/localeManager';
 
 export const DEFAULT_OVERLAY_CLASS = 'ag-chart-overlay';
 export const DEFAULT_OVERLAY_DARK_CLASS = 'ag-chart-dark-overlay';
