+++ packages/ag-charts-community/src/dom/domManager.ts
@@ -1,10 +1,10 @@
 import type { AgIconName } from 'ag-charts-types';
 
-import { BBox } from '../../scene/bbox';
-import STYLES from '../../styles.css';
-import { createElement, getDocument, getWindow } from '../../util/dom';
-import { type Size, SizeMonitor } from '../../util/sizeMonitor';
-import { BaseManager } from '../baseManager';
+import { BBox } from '../scene/bbox';
+import STYLES from '../styles.css';
+import { createElement, getDocument, getWindow } from '../util/dom';
+import { type Size, SizeMonitor } from '../util/sizeMonitor';
+import { BaseManager } from '../chart/baseManager'; // TODO move to utils
 import BASE_DOM from './domLayout.html';
 
 const CANVAS_CENTER_CLASS = 'canvas-center';
