+++ packages/ag-charts-community/src/dom/domStyles.css
