+++ packages/ag-charts-community/src/dom/focusStyles.css
