+++ packages/ag-charts-community/src/dom/icons.css
