+++ packages/ag-charts-community/src/locale/defaultMessageFormatter.ts
@@ -1,6 +1,6 @@
 import type { Formatter, MessageFormatterParams } from 'ag-charts-types';
 
-import { Logger } from '../../util/logger';
+import { Logger } from '../util/logger';
 
 const messageRegExp = /\$\{(\w+)}(?:\[(\w+)])?/gi;
