+++ packages/ag-charts-community/src/module/moduleContext.ts
@@ -3,9 +3,9 @@ import type { AnnotationManager } from '../chart/annotation/annotationManager';
 import type { AxisManager } from '../chart/axis/axisManager';
 import type { ChartService } from '../chart/chartService';
 import type { DataService } from '../chart/data/dataService';
-import type { DOMManager } from '../chart/dom/domManager';
-import type { FocusIndicator } from '../chart/dom/focusIndicator';
-import type { ProxyInteractionService } from '../chart/dom/proxyInteractionService';
+import type { DOMManager } from '../dom/domManager';
+import type { FocusIndicator } from '../dom/focusIndicator';
+import type { ProxyInteractionService } from '../dom/proxyInteractionService';
 import type { AnimationManager } from '../chart/interaction/animationManager';
 import type { AriaAnnouncementService } from '../chart/interaction/ariaAnnouncementServices';
 import type { ChartEventManager } from '../chart/interaction/chartEventManager';
@@ -21,7 +21,7 @@ import type { ToolbarManager } from '../chart/interaction/toolbarManager';
 import type { TooltipManager } from '../chart/interaction/tooltipManager';
 import type { ZoomManager } from '../chart/interaction/zoomManager';
 import type { LayoutManager } from '../chart/layout/layoutManager';
-import type { LocaleManager } from '../chart/locale/localeManager';
+import type { LocaleManager } from '../locale/localeManager';
 import type { SeriesStateManager } from '../chart/series/seriesStateManager';
 import type { UpdateService } from '../chart/updateService';
 import type { Scene } from '../scene/scene';
